Add comprehensive tests for CPU instructions and memory operations
- Implement tests for the LDI instruction covering all shift cases and immediate loading.
- Add tests for memory operations including LOAD/STORE, BYTE_LOAD/STORE, SHORT_LOAD/STORE, and WORD_LOAD/STORE.
- Introduce tests for special instructions such as STOP, LSR, SSR, and IRET.
- Create a shared test harness to streamline test execution and assertions.
- Update linker tests to reflect changes in JUMP.Z instruction encoding.

+++ disassembler/disassembler.hpp
@@ -21,7 +21,9 @@ struct DisassembledInstruction {
     uint8_t rs1;                // bits[7:4]: base register (also used by format 11)
 
     // Format 01 only
-    int8_t  pc_rel;             // bits[9:4]: signed 6-bit; byte offset = pc_rel * 2
+    // For non-JUMP opcodes: 6-bit signed offset from bits[9:4]; bits[3:0] are Rd.
+    // For JUMP.* opcodes (11–15): 10-bit signed offset from bits[9:0]; Rd is always R15.
+    int16_t pc_rel;             // signed offset in instruction units; byte offset = pc_rel * 2
     uint16_t effective_address; // address + 2 + (pc_rel * 2)
 
     // Format 10 only
